+++ test/editor/serialize-test.js
@@ -32,9 +32,9 @@ describe('editor/serialize', function() {
         const html = htmlSerializeIfNeeded(model, {});
         expect(html).toBe("<a href=\"https://matrix.to/#/#room:hs.tld\">#room:hs.tld</a>");
     });
-    it('@room pill turns message into html', function() {
+    it('@all pill turns message into html', function() {
         const pc = createPartCreator();
-        const model = new EditorModel([pc.atRoomPill("@room")]);
+        const model = new EditorModel([pc.atRoomPill("@all")]);
         const html = htmlSerializeIfNeeded(model, {});
         expect(html).toBeFalsy();
     });